Add experimental GC context and object scanner APIs
This commit introduces two experimental APIs to the GC. Both APIs are
only supported under cmake, and are permanantly disabled in autotools.

The first is a high level wrapper for initializing and tearing down the
GC. To enable this API, use the flag OMR_GC_EXPERIMENTAL_CONTEXT.
This code is grouped under the gc/api directory.

In the experimental context APIs, the  GC is structured into three
tiers:

1. OMR::Runtime - RAII-style process-wide initialization. Creating a
   Runtime will bring  up the thread library, port library, and
   OMR_Runtime. There should only be one instance of the runtime in a
   program.
2. OMR::GC::System - A managed heap. Provides it's own shared-access
   mechanism. Typically used on a per-VM basis.
3. OMR::GC::Context - per-thread context attached to a GC::System. Most
   GC APIs require a GC::Context. Having a context implies shared heap
   accss. Contexts can be upgraded to exclusive contexts.

The implementation of this API is placed under gc/api. New public
headers are grouped under gc/api/include/OMR/GC/*. They are meant to be
included as #include <OMR/GC/Context.hpp>.

GC::System  implements it's own shared access mechanism.  Implementing
it's own heap access mechanism requires that some GC glue is actually
provided by the GC, such as the EnvironmentLanguageInterface.

To support this notion of gc-provided glue, this PR rearranges the
example's glue headers into include and include_extended.
include_extended is only added to the ipath if
OMR_GC_EXPERIMENTAL_CONTEXT is off, to ensure there is only ever one
version of each glue header available.

The second API is a new piece of glue, the OMRClient::GC::ObjectScanner.
This is an attempt to write an object scanner API that is simpler for
clients to implement, while  supporting custom reference encodings. The
new glue APIs are placed under the OMRClient namespace. Public headers
are included as <OMRClient/GC/ObjectScanner.hpp>.

The new object scanner API

The experimental object scanner makes compressed object pointers a
concern of the application. The long term goal is to remove the
compressed references flag from OMR entirely. Through utility classes
provided by OMR, implementing compressed references should be trivial
for applications to implement.

The object scanner, if successful, will deprecate:
* fomrobject_t
* GC_SlotObject

Signed-off-by: Robert Young <rwy0717@gmail.com>